Move DateCollection class to avoid circular dependency.
